backend/db_api/pyproject.toml
[tool.poetry]
name = "db_api"
version = "0.1.0"
description = ""
authors = ["engineers <engineers@bogeyians.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
FastAPI-SQLAlchemy = "^0.2.1"
pydantic = "^1.10.2"
alembic = "^1.8.1"
psycopg2 = "^2.9.3"
uvicorn = "^0.18.3"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
